fix(review): address second wave of pr #1244 review threads (~34)
Conditioner inclusion (11 files)
- FlowMap, OSDS, SD3Flash, FluxInpainting, HDPainter, PowerPaint,
  PromptToPrompt, TurboEdit, MotionDiffuse, DiffBIR, HiDream:
  ParameterCount now adds the conditioner when it implements
  IParameterizable, matching GetParameterChunks's three-component
  stream

VAE inclusion in ControlARModel + PixArtModel
- both classes own and expose _vae but were excluding it from
  ParameterCount/GetParameters/SetParameters; bring them in line
  with the rest of the LatentDiffusion family. ControlARModel
  GetParameters/SetParameters now use a long expectedTotal and
  validate length up front.

Materialization-just-to-count (38 more files)
- replace `_unet.GetParameters().Length` (and `_baseUNet`)
  with `checked((int)_unet.ParameterCount)`. The U-Net buffer is
  walked just to read .Length, defeating the long-safe migration.
  This catches the call sites the prior `_vae.GetParameters().Length`
  fix missed because the variable name differed.

Length validation in SetParameters
- ControlNetPlusPlusModel, ControlNetTileModel: validate the full
  vector length up front against the long expectedTotal before
  slicing, so undersized inputs fail with a clear message.

Bare (int) casts -> checked() casts
- ControlNetUnionProModel encoder loop, IPAdapterModel transformer
  layer loop and token-expansion read: silent overflow could land
  encCount/count negative; checked() throws cleanly.

Heavy-lift Clone refactor
- FluxSchnellModel: replace flat GetParameters/SetParameters round
  trip in Clone() with field-by-field copies. ~12B FLUX-scale
  weights cross int.MaxValue and the flat path is unsupportable.

ProgressiveNeuralNetworks stats
- GetNetworkStats sums per-column ParameterCount as long, so
  TotalParameters/FrozenParameters/ActiveParameters preserve aggregate
  counts beyond int.MaxValue.

PASDModel GetParameters arithmetic
- widen pp.Length + vp.Length to long before the Vector<T> ctor;
  checked() narrows at the boundary so foundation-scale combinations
  fail fast.

+++ src/ContinualLearning/ProgressiveNeuralNetworks.cs
@@ -265,12 +265,15 @@ public class ProgressiveNeuralNetworks<T> : IContinualLearningStrategy<T>
     /// <returns>Dictionary with network statistics.</returns>
     public Dictionary<string, object> GetNetworkStats()
     {
+        // Sum into long to preserve aggregate counts beyond int.MaxValue
+        // (each column already fits in int by Vector<T> indexability,
+        // but the sum across columns doesn't).
         return new Dictionary<string, object>
         {
             ["ColumnCount"] = _columns.Count,
-            ["TotalParameters"] = _columns.Sum(c => (int)c.ParameterCount),
-            ["FrozenParameters"] = _columns.Where(c => c.IsFrozen).Sum(c => (int)c.ParameterCount),
-            ["ActiveParameters"] = _columns.Where(c => !c.IsFrozen).Sum(c => (int)c.ParameterCount),
+            ["TotalParameters"] = _columns.Sum(c => (long)c.ParameterCount),
+            ["FrozenParameters"] = _columns.Where(c => c.IsFrozen).Sum(c => (long)c.ParameterCount),
+            ["ActiveParameters"] = _columns.Where(c => !c.IsFrozen).Sum(c => (long)c.ParameterCount),
             ["UseLateralConnections"] = _useLateralConnections,
             ["TaskIds"] = _columns.Select(c => c.TaskId).ToList()
         };
